==> notebooks/03_preprocessing.ipynb <==
# Preprocessing Step (Etapa de Pré-Processamento)

## Table of Contents
* [Packages](#1)
* [Preprocessing](#2)

<a name="1"></a>
## Packages (Pacotes)

import pandas as pd
import string
import re
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
import tensorflow as tf
from tensorflow.keras.layers import TextVectorization

<a name="2"></a>
## Preprocessing (Pré-processamento)

comics_data = pd.read_csv('../data/raw/comics_corpus.csv').drop_duplicates()
char_data = pd.read_csv('../data/raw/char_corpus.csv')
comics_data.head()

char_data.head()

comics_data['description'].str.contains('!').sum()

MAX_TOKENS = None
MAX_LEN = max([len(text) for text in comics_data['description']])
MAX_LEN

def rnn_preprocess(corpus):
    """
    
    """
    stem = PorterStemmer()
    stopwords_en = stopwords.words('english')
    
    corpus = re.sub(f'r[{re.escape(string.punctuation)}]', '.', corpus)
    corpus = re.sub(f'r{stopwords_en}', '', corpus)
    corpus = re.sub(r'\s+', ' ', corpus)
    corpus = re.sub(r'\W+', ' ', corpus).strip().lower()   
    
    return corpus

comics_data['description'].map(rnn_preprocess)

comics_data['description'] = comics_data['description'].map(rnn_preprocess)
comics_data.head()

comics_data['description'].str.contains('!').sum()

def rnn_vectorization(corpus, max_tokens=None):
    """
    
    """
    sentence_vec = TextVectorization(
        standardize='lower_and_strip_punctuation',
        max_tokens=max_tokens,
        ragged=True
    )
    sentence_vec.adapt(corpus)
    vocabulary = sentence_vec.get_vocabulary(include_special_tokens=True)

    return sentence_vec, vocabulary

sentence_vec, vocabulary = rnn_vectorization(comics_data['description'], max_tokens=None)
len(vocabulary)

X_vec = sentence_vec(comics_data['description'])

def padding(corpus, maxlen=None, shuffle_buffer_size=1000, prefetch_buffer_size=tf.data.AUTOTUNE, batch_size=128):
    """
    
    """
    padded_corpus = tf.keras.utils.pad_sequences(
        corpus.numpy(),
        maxlen=maxlen,
        truncating='post',
        padding='post'
    )
    padded_corpus = (
        tf.data.Dataset.from_tensor_slices(padded_corpus)
        .cache()
        .shuffle(shuffle_buffer_size)
        .prefetch(prefetch_buffer_size)
        .batch(batch_size)
    )

    return padded_corpus

X_pad = padding(X_vec, maxlen=MAX_LEN)

next(iter(X_pad.take(1)))

y_dataset = tf.data.Dataset.from_tensor_slices(comics_data['y'])
comics_dataset = tf.data.Dataset.zip(X_pad, y_dataset)
next(iter(comics_dataset.take(1)))

punctuations = string.punctuation
stopwords_en = stopwords.words('english')
print(f'Punctuations: {punctuations}\n\nStopwords: {stopwords_en}')

re.escape(' '.join(stopwords_en))
